+++ lib/bond/compiler/annotated_function/clause.ex
@@ -6,16 +6,26 @@ defmodule Bond.Compiler.AnnotatedFunction.Clause do
 
   alias Bond.Compiler.FunctionDefinition
 
-  defstruct [:env, :params, :guards, :body]
+  defstruct [:env, :params, :guards, :body, :warn_skipped_invariants_override]
 
   @type t :: %__MODULE__{
           env: Macro.Env.t() | nil,
           params: list() | nil,
           guards: list() | nil,
-          body: keyword() | nil
+          body: keyword() | nil,
+          warn_skipped_invariants_override: nil | boolean()
         }
 
   def new(%FunctionDefinition{} = function_def) do
-    struct(__MODULE__, Map.take(function_def, [:env, :params, :guards, :body]))
+    struct(
+      __MODULE__,
+      Map.take(function_def, [
+        :env,
+        :params,
+        :guards,
+        :body,
+        :warn_skipped_invariants_override
+      ])
+    )
   end
 end